+++ tests/bunit.core.tests/ComponentFactories/GenericComponentFactoryTest.cs
@@ -11,11 +11,11 @@ namespace Bunit.ComponentFactories
 {
 	public class GenericComponentFactoryTest : TestContext
 	{
-		[Fact(DisplayName = "UseFor throws when factories is null")]
+		[Fact(DisplayName = "Add throws when factories is null")]
 		public void Test001()
 			=> Should.Throw<ArgumentNullException>(() => ComponentFactoryCollectionExtensions.Add<Simple1, FakeSimple1>(factories: default));
 
-		[Fact(DisplayName = "UseFor<TComponent, TReplacementComponent> replaces components of type TComponent with TReplacementComponent")]
+		[Fact(DisplayName = "Add<TComponent, TReplacementComponent> replaces components of type TComponent with TReplacementComponent")]
 		public void Test002()
 		{
 			ComponentFactories.Add<Simple1, FakeSimple1>();